replace old PLDCorrector
