+++ lightkurve/correctors/pldcorrector.py
@@ -1,16 +1,65 @@
-"""Defines PLDCorrector
+"""Defines PLDCorrector.
+
+This module requires 3 optional dependencies (theano, pymc3, exoplanet) for
+PLD correction to work.  One additional dependency (fbpca) is not required
+but will speed up the computation if installed.
+
+TODO Now
+--------
+* [DONE] Implement diagnostic plots, `corr.plot_design_matrix()` and `corr.plot_diagnostics()`.
+* Port the existing suite of PLDCorrector tests.
+* [kinda done] Fix the units of corrected_flux, corrected_flux_without_star, etc.
+  In Lightkurve we currently add back the mean of the model.
+* [DONE] Document the meaning of logs2, logsigma (amplitude of variability),
+  and logrho (timescale of variability), and make sure their
+  prior values are configurable.
+* [DONE] Add description of new PyMC3 implimentation to docstring
+* [DONE] Rename star model
+
+TODO Before release
+-------------------
+* Rename PyMCPLDCorrector as PLDCorrector.
+* The design matrix can be improved by rejecting pixels which are saturated,
+  and including the collapsed sums of their CCD columns instead.
+* Add pymc & exoplanet & theano to Lightkurve's dependencies or treat it properly
+  as an optional import. [Geert]
+* Set PyMC verbosity to match the lightkurve.log.level. [Geert]
+
+Future enhancements
+-------------------
+* In `create_first_order_matrix`, we need to explain why we are dividing each
+  cadence by the sum of all pixels. We think this is explained in the EVEREST paper,
+  i.e. is is likely because PLD optimizes towards a flat light curve. We need
+  to find the right words to explain this and refer to papers where possible.
+* It is not clear whether the inclusion of a column vector of ones in the
+  design matrix is necessary for numerical stability.
 """
 from __future__ import division, print_function
 
 import logging
-import warnings
 from itertools import combinations_with_replacement as multichoose
 
 import numpy as np
+import matplotlib.pyplot as plt
+
+from .. import LightCurve
+
+# Optional dependencies
+try:
+    import pymc3 as pm
+    import exoplanet as xo
+    import theano.tensor as tt
+    tt.config.exception_verbosity='high'
+except ImportError:
+    # Fail quietly here so we don't break `import lightkurve`.
+    # We will raise a user-friendly ImportError inside PLDCorrector.__init__().
+    pass
+
+from .. import MPLSTYLE
 
 log = logging.getLogger(__name__)
 
-__all__ = ['PLDCorrector']
+__all__ = ['PyMCPLDCorrector']
 
 
 class PLDCorrector(object):
@@ -26,13 +75,13 @@ class PLDCorrector(object):
         of PLD that is less sophisticated than EVEREST, but is suitable
         for quick-look analyses and detrending experiments.
 
-        Our simple implementation of PLD is performed by first calculating the
-        noise model for each cadence in time. This function goes up to arbitrary
+        Our implementation of PLD is performed by first calculating the noise
+        model for each cadence in time. This function goes up to arbitrary
         order, and is represented by
 
         .. math::
 
-            m_i = \alpha + \beta t_i + \gamma t_i^2 + \sum_l a_l \frac{f_{il}}{\sum_k f_{ik}} + \sum_l \sum_m b_{lm} \frac{f_{il}f_{im}}{\left( \sum_k f_{ik} \right)^2} + ...
+            m_i = \sum_l a_l \frac{f_{il}}{\sum_k f_{ik}} + \sum_l \sum_m b_{lm} \frac{f_{il}f_{im}}{\left( \sum_k f_{ik} \right)^2} + ...
         where
 
           - :math:`m_i` is the noise model at time :math:`t_i`
@@ -40,13 +89,20 @@ class PLDCorrector(object):
           - :math:`a_l` is the first-order PLD coefficient on the linear term
           - :math:`b_{lm}` is the second-order PLD coefficient on the :math:`l^\text{th}`,
             :math:`m^\text{th}` pixel pair
-          - :math:`\alpha`, :math:`\beta`, and :math:`\gamma` are the
-            Gaussian Process terms applied to capture long-period variability.
 
         We perform Principal Component Analysis (PCA) to reduce the number of
         vectors in our final model to limit the set to best capture instrumental
         noise. With a PCA-reduced set of vectors, we can construct a design matrix
         containing fractional pixel fluxes.
+
+        To capture long-term variability, we simultaneously fit a Gaussian Process
+        model ([5]_) to the underlying stellar signal. We use the gradient-based
+        probabilistic modeling toolkit [6]_ to optimize the GP hyperparameters and
+        solve for the motion model.
+
+        To robustly estimate errors on our parameter estimates and output flux
+        values, we optionally sample the output model with [7]_ and infer errors from
+        the posterior distribution.
 
         To solve for the PLD model, we need to minimize the difference squared
 
@@ -60,19 +116,6 @@ class PLDCorrector(object):
 
             \frac{\partial \chi^2}{\partial a_l} = 0.
 
-    Examples
-    --------
-    Download the pixel data for GJ 9827 and obtain a PLD-corrected light curve:
-
-    >>> import lightkurve as lk
-    >>> tpf = lk.search_targetpixelfile("GJ9827").download() # doctest: +SKIP
-    >>> corrector = lk.PLDCorrector(tpf) # doctest: +SKIP
-    >>> lc = corrector.correct() # doctest: +SKIP
-    >>> lc.plot() # doctest: +SKIP
-
-    However, the above example will over-fit the small transits!
-    It is necessary to mask the transits using `corrector.correct(cadence_mask=...)`.
-
     References
     ----------
     .. [1] Deming et al. (2015), ads:2015ApJ...805..132D.
@@ -82,199 +125,470 @@ class PLDCorrector(object):
     .. [3] Luger et al. (2018), ads:2018AJ....156...99L
         (arXiv:1702.05488)
     .. [4] EVEREST pipeline webpage, https://rodluger.github.io/everest
+    .. [5] Celerite documentation, https://celerite.readthedocs.io/en/stable/
+    .. [6] Exoplanet documentation, https://exoplanet.readthedocs.io/en/stable/
+    .. [7] PyMC3 documentation, https://docs.pymc.io
+
+    Parameters
+    ----------
+    tpf : TargetPixelFile
+    aperture_mask : 2D boolean array or str
+    pld_aperture_mask : 2D boolean array or str
+
+    Raises
+    ------
+    ImportError : if one of Lightkurve's optional dependencies required by
+        this class are missing.
     """
-    def __init__(self, tpf):
-        self.tpf = tpf
+    def __init__(self, tpf, aperture_mask=None, pld_aperture_mask=None):
+        # Ensure the optional dependencies requires by this class are installed
+        success, messages = self._check_optional_dependencies()
+        if not success:
+            [log.error(message) for message in messages]
+            raise ImportError("\n".join(messages))
 
-        self.flux = tpf.flux
-        self.flux_err = tpf.flux_err
-        self.time = tpf.time
+        # Input validation: parse the aperture masks to accept strings etc.
+        self.aperture_mask = tpf._parse_aperture_mask(aperture_mask)
+        self.pld_aperture_mask = tpf._parse_aperture_mask(pld_aperture_mask)
+        # Generate raw flux light curve from desired pixels
+        raw_lc = tpf.to_lightcurve(aperture_mask=self.aperture_mask)
+        # It is critical to remove all NaNs or the linear algebra below will crash
+        self.raw_lc, self.nan_mask = raw_lc.remove_nans(return_mask=True)
+        self.tpf = tpf[~self.nan_mask]
+        # For user-friendliness, store most recent solutions
+        self.most_recent_solution = None
+        self.most_recent_trace = None
 
-    def correct(self, aperture_mask=None, cadence_mask=None, gp_timescale=30,
-                use_gp=True, pld_order=2, n_pca_terms=10, pld_aperture_mask=None):
-        r"""Returns a PLD systematics-corrected LightCurve.
-
-        Parameters
-        ----------
-        aperture_mask : array-like, 'pipeline', 'all', 'threshold', or None
-            A boolean array describing the aperture such that `True` means
-            that the pixel will be used to generate the raw flux light curve.
-            If `None` or 'all' are passed, all pixels will be used.
-            If 'pipeline' is passed, the mask suggested by the official pipeline
-            will be returned.
-            If 'threshold' is passed, all pixels brighter than 3-sigma above
-            the median flux will be used.
-        cadence_mask : array-like
-            A mask that will be applied to the cadences prior to constructing
-            the detrending model. For example, you can pass a boolean array
-            of length `n_cadences` where `True` means that the cadence will be
-            included in the noise model. You may also pass an array of indices.
-            This option enables signals of interest (e.g. planet transits)
-            to be excluded from the noise model, which will prevent over-fitting.
-            By default, no cadences will be masked.
-        gp_timescale : float
-            Gaussian Process time scale length term (`tau`) used to define
-            length of fit variability in days.
-        use_gp : boolean
-            Option to turn GP fitting on or off.  You would typically only set
-            this to False to speed up the correction (at the cost of precision),
-            or if you suspect the presence of systematic noise at long timescales.
-        pld_order : int
-            The order of Pixel Level De-correlation to be performed. First order
-            (`n=1`) uses only the pixel fluxes to construct the design matrix.
-            Higher order populates the design matrix with columns constructed
-            from the products of pixel fluxes.
-        n_pca_terms : int
-            Number of terms added to the design matrix from each order of PLD
-            when performing Principal Component Analysis for models higher than
-            first order. Increasing this value may provide higher precision at
-            the expense of computational time.
-        pld_aperture_mask : array-like, 'pipeline', 'all', 'threshold', or None
-            A boolean array describing the aperture such that `True` means
-            that the pixel will be used when selecting the PLD basis vectors.
-            If `None` or `all` are passed in, all pixels will be used.
-            If 'pipeline' is passed, the mask suggested by the official pipeline
-            will be returned.
-            If 'threshold' is passed, all pixels brighter than 3-sigma above
-            the median flux will be used.
+    def _check_optional_dependencies(self):
+        """Emits a user-friendly error message if one of Lightkurve's
+        optional dependencies which are required for PLDCorrector are missing.
 
         Returns
         -------
-        corrected_lightcurve : `~lightkurve.lightcurve.LightCurve`
-            Returns a corrected lightcurve object. Depending on the input, the
-            returned object will be a `KeplerLightCurve`, `TessLightCurve`, or
-            general `LightCurve` object.
+        success : bool
+            True if all optional dependencies are available, False otherwise.
+        message : list of str
+            User-friendly error message if success == False.
         """
-        if use_gp:
-            # Verify optional dependency
-            try:
-                import celerite
-            except ImportError:
-                log.error("PLD uses the `celerite` Python package. "
-                          "See the installation instructions at "
-                          "https://docs.lightkurve.org/about/install.html. "
-                          "`use_gp` has been set to `False`.")
-                use_gp = False
+        success, messages = True, []
 
-        # Parse the aperture mask to accept strings etc.
-        aperture = self.tpf._parse_aperture_mask(aperture_mask)
+        try:
+            import pymc3 as pm
+        except ImportError as e:
+            success = False
+            messages.append("PLDCorrector requires pymc3 to be installed (`pip install pymc3`).")
 
-        # generate flux light curve from desired pixels
-        lc = self.tpf.to_lightcurve(aperture_mask=aperture)
-        rawflux = lc.flux
-        rawflux_err = lc.flux_err
+        try:
+            import exoplanet as xo
+        except ImportError as e:
+            success = False
+            messages.append("PLDCorrector requires exoplanet to be installed (`pip install exoplanet`).")
 
-        # create nan mask
-        nanmask = np.isfinite(self.time)
-        nanmask &= np.isfinite(rawflux)
-        nanmask &= np.isfinite(rawflux_err)
-        nanmask &= np.abs(rawflux_err) > 1e-12
+        try:
+            import theano.tensor as tt
+        except ImportError as e:
+            success = False
+            message.append("PLDCorrector requires theano to be installed (`pip install theano`).")
 
-        # mask out nan values
-        rawflux = rawflux[nanmask]
-        rawflux_err = rawflux_err[nanmask]
-        self.flux = self.flux[nanmask]
-        self.flux_err = self.flux_err[nanmask]
-        self.time = self.time[nanmask]
+        return success, messages
 
-        # parse the PLD aperture mask
-        pld_pixel_mask = self.tpf._parse_aperture_mask(pld_aperture_mask)
+    def create_first_order_matrix(self):
+        """Returns a matrix which encodes the fractional pixel fluxes as a function
+        of cadence (row) and pixel (column). As such, the method returns a
+        2D matrix with shape (n_cadences, n_pixels_in_pld_mask).
 
-        # find pixel bounds of aperture on tpf
-        xmin, xmax = min(np.where(pld_pixel_mask)[0]),  max(np.where(pld_pixel_mask)[0])
-        ymin, ymax = min(np.where(pld_pixel_mask)[1]),  max(np.where(pld_pixel_mask)[1])
+        This matrix will form the basis of the PLD regressor design matrix
+        and is often called the first order component.
 
-        # crop data cube to include only desired pixels
-        # this is required for superstamps to ensure matrix is invertable
-        flux_crop = self.flux[:, xmin:xmax+1, ymin:ymax+1]
-        flux_err_crop = self.flux_err[:, xmin:xmax+1, ymin:ymax+1]
-        aperture_crop = pld_pixel_mask[xmin:xmax+1, ymin:ymax+1]
+        The matrix returned is guaranteed to be free of NaN values.
 
-        # calculate errors (ignore warnings related to zero or negative errors)
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", RuntimeWarning)
-            flux_err = np.nansum(flux_err_crop[:, aperture_crop]**2, axis=1)**0.5
+        Returns
+        -------
+        matrix : numpy array
+            First order PLD design matrix.
+        """
+        # Re-arrange the cube of flux values observed in a user-specified mask
+        # into a 2D matrix of shape (n_cadences, n_pixels_in_mask).
+        # Note that Theano appears to require 64-bit floats.
+        matrix = np.asarray(self.tpf.flux[:, self.pld_aperture_mask], np.float64)
+        assert matrix.shape == (len(self.raw_lc.time), self.pld_aperture_mask.sum())
+        # Remove all NaN or Inf values
+        matrix = matrix[:, np.isfinite(matrix).all(axis=0)]
+        # To ensure that each column contains the fractional pixel flux,
+        # we divide by the sum of all pixels in the same cadence.
+        # The reason for doing this is described in Section 2 of Luger et al. (2016).
+        matrix = matrix / np.sum(matrix, axis=-1)[:, None]
+        # If we return matrix at this point, theano will raise a "dimension mismatch".
+        # The origin of this bug is not understood, but copying the matrix
+        # into a new one as shown below circumvents it:
+        result = np.empty((matrix.shape[0], matrix.shape[1]))
+        result[:, :] = matrix[:, :]
 
-        # first order PLD design matrix
-        pld_flux = flux_crop[:, aperture_crop]
-        f1 = np.reshape(pld_flux, (len(pld_flux), -1))
-        X1 = f1 / np.nansum(pld_flux, axis=-1)[:, None]
-        # No NaN pixels
-        X1 = X1[:, np.isfinite(X1).all(axis=0)]
+        return result
 
-        # higher order PLD design matrices
-        X_sections = [np.ones((len(flux_crop), 1)), X1]
-        for i in range(2, pld_order+1):
-            f2 = np.product(list(multichoose(X1.T, pld_order)), axis=1).T
-            try:
-                # We use an optional dependency for very fast PCA (fbpca).
-                # If the import fails we will fall back on using the slower `np.linalg.svd`
-                from fbpca import pca
-                components, _, _ = pca(f2, n_pca_terms)
-            except ImportError:
-                log.error("PLD uses the `fbpca` package. You can pip install "
-                          "with `pip install fbpca`. Using `np.linalg.svd` "
-                          "instead.")
-                components, _, _ = np.linalg.svd(f2)
-            X_n = components[:, :n_pca_terms]
-            X_sections.append(X_n)
+    def create_design_matrix(self, pld_order=1, n_pca_terms=10, include_column_of_ones=False):
+        """Returns a matrix designed to contain suitable regressors for the
+        systematics noise model.
 
-        # Create the design matrix X by stacking X1 and higher order components, and
-        # adding a column vector of 1s for numerical stability (see Luger et al.).
-        # X has shape (n_components_first + n_components_higher_order + 1, n_cadences)
-        X = np.concatenate(X_sections, axis=1)
+        The design matrix contains one row for each cadence (i.e. moment in time)
+        and one column for each regressor that we wish to use to predict the
+        systematic noise in a given cadence.
 
-        # set default transit mask
+        The columns (i.e. regressors) included in the design matrix are:
+        * One column for each pixel in the PLD aperture mask.  Each column
+          contains the flux values observed by that pixel over time.  This is
+          also known as the first order component.
+        * Columns derived from the products of all combinations of pixel values
+          in the aperture mask. However, rather than including a column for each
+          combination, we perform dimensionality reduction (PCA) and include a
+          smaller number of PCA terms, i.e. the number of columns is
+          n_pca_terms*(pld_order-1).  This is also known as the higher order
+          components.
+        * Optionally, a single column of ones for numerical stability.
+
+        Thus, the shape of the design matrix will be
+        (n_cadences, n_pld_mask_pixels + n_pca_terms*(pld_order-1) + include_column_of_ones)
+
+        Returns
+        -------
+        design_matrix : 2D numpy array
+            See description above.
+        """
+        # We use an optional dependency for very fast PCA (fbpca), but if the
+        # import fails we will fall back on using the slower `np.linalg.svd`.
+        use_fbpca = True
+        try:
+            from fbpca import pca
+        except ImportError:
+            use_fbpca = False
+            log.warning("PLD systematics correction will run faster if the "
+                        "optional `fbpca` package is installed "
+                        "(`pip install fbpca`).")
+
+        matrix_sections = []  # list to hold the design matrix components
+        first_order_matrix = self.create_first_order_matrix()
+
+        # Input validation: n_pca_terms cannot be larger than the number of cadences
+        n_cadences = len(first_order_matrix)
+        if n_pca_terms > n_cadences:
+            log.warning("`n_pca_terms` ({}) cannot be larger than the number of cadences ({});"
+                        "using n_pca_terms={}".format(n_pca_terms, n_cadences, n_cadences))
+            n_pca_terms = n_cadences
+
+        # The original EVEREST paper includes a column vector of ones in the
+        # design matrix to improve the numerical stability (see Luger et al.);
+        # it is unclear whether this is necessary, so this is an optional step
+        # for now.
+        if include_column_of_ones:
+            matrix_sections.append([np.ones((len(first_order_matrix), 1))])
+
+        # Add the first order matrix
+        matrix_sections.append(first_order_matrix)
+
+        # Add the higher order PLD design matrix columns
+        for order in range(2, pld_order + 1):
+            # Take the product of all combinations of pixels; order=2 will
+            # multiply all pairs of pixels, order=3 will multiple triples, etc.
+            matrix = np.product(list(multichoose(first_order_matrix.T, order)), axis=1).T
+            # This product matrix becomes very big very quickly, so we reduce
+            # its dimensionality using PCA.
+            if use_fbpca:  # fast mode
+                components, _, _ = pca(matrix, n_pca_terms)
+            else:  # slow mode
+                components, _, _ = np.linalg.svd(matrix)
+            section = components[:, :n_pca_terms]
+            matrix_sections.append(section)
+
+        return np.concatenate(matrix_sections, axis=1)
+
+    def create_pymc_model(self, design_matrix=None, cadence_mask=None, gp_timescale_prior=150, **kwargs):
+        r"""Returns a PYMC3 model.
+
+        Parameters
+        ----------
+        design_matrix : np.ndarray
+            Matrix of shape (n_cadences, n_regressors) used to create the
+            motion model.
+        cadence_mask : np.ndarray
+            Boolean array to mask cadences. Cadences that are False will be excluded
+            from the model fit
+        gp_timescale_prior : int
+            The parameter `rho` in the definition of the Matern-3/2 kernel, which
+            influences the timescale of variability fit by the Gaussian Process.
+            For more information, see [1]
+
+        Returns
+        -------
+        model : pymc3.model.Model
+            A pymc3 model
+
+        References
+        ----------
+        .. [1] the `celerite` documentation https://celerite.readthedocs.io
+        """
+        if design_matrix is None:
+            design_matrix = self.create_design_matrix(**kwargs)
         if cadence_mask is None:
-            cadence_mask = np.ones_like(lc.time, dtype=bool)
-        M = lambda x: x[cadence_mask[nanmask]]
+            cadence_mask = np.ones(len(self.raw_lc.time), dtype=bool)
 
-        # mask transits in design matrix
-        MX = M(X)
+        # Theano raises an `AsTensorError` ("Cannot convert to TensorType")
+        # if the floats are not in 64-bit format, so we convert them here:
+        lc_time = np.asarray(self.raw_lc.time, np.float64)
+        lc_flux = np.asarray(self.raw_lc.flux, np.float64)
+        lc_flux_err = np.asarray(self.raw_lc.flux_err, np.float64)
 
-        if use_gp:
-            # We use a Gaussian Process to model the long term trend.
-            # We do this by estimating the long term trend y by applying the
-            # preliminary PLD model defined above and subtracting it from the raw light curve.
-            # The "in transit" cadences are masked out in this step to prevent the
-            # long term approximation from over-fitting the transits.
-            XTX = np.dot(MX.T, MX)
-            XTX[np.diag_indices_from(XTX)] += 1e-8
-            XTy = np.dot(MX.T, M(rawflux))
-            y = M(rawflux) - np.dot(MX, np.linalg.solve(XTX, XTy))
+        # Covariance matrix diagonal
+        diag = lc_flux_err**2
 
-            # Estimate the amplitude parameter of a Matern-3/2 kernel GP
-            # by computing the standard deviation of y.
-            amp = np.nanstd(y)
-            tau = gp_timescale  # tau is a user-defined parameter
-            # set up gaussian process using celerite
-            # we use a Matern-3/2 kernel for its flexibility and non-periodicity
-            kernel = celerite.terms.Matern32Term(np.log(amp), np.log(tau))
-            gp = celerite.GP(kernel)
-            gp.compute(M(self.time), M(rawflux_err))
+        # The cadence mask is applied by inflating the uncertainties in the covariance matrix;
+        # this is because celerite will run much faster if it is able to predict
+        # data for cadences that have been fed into the model.
+        diag[~cadence_mask] += 1e12
 
-            # compute the coefficients C on the basis vectors;
-            # the PLD design matrix will be dotted with C to solve for the noise model.
-            A = np.dot(MX.T, gp.apply_inverse(MX))
-            B = np.dot(MX.T, gp.apply_inverse(M(rawflux)[:, None])[:, 0])
+        with pm.Model() as model:
+            # Create a Gaussian Process to model the long-term stellar variability
+            # log(sigma) is the amplitude of variability, estimated from the raw flux scatter
+            logsigma = pm.Normal("logsigma", mu=np.log(np.std(lc_flux)), sd=4)
+            # log(rho) is the timescale of variability with a user-defined prior
+            logrho = pm.Normal("logrho", mu=np.log(gp_timescale_prior), sd=4)
+            # Add a jitter term
+            logs2 = pm.Normal("logs2", mu=np.log(np.var(lc_flux)), sd=4)
+            kernel = xo.gp.terms.Matern32Term(log_sigma=logsigma, log_rho=logrho)
 
+            # Store the GP and cadence mask to aid debugging
+            model.gp = xo.gp.GP(kernel, lc_time, diag + tt.exp(logs2))
+            self.cadence_mask = cadence_mask
+
+            # The motion model regresses against the design matrix
+            A = tt.dot(design_matrix.T, model.gp.apply_inverse(design_matrix))
+            B = tt.dot(design_matrix.T, model.gp.apply_inverse(lc_flux[:, None]))
+            weights = tt.slinalg.solve(A, B)
+            motion_model = pm.Deterministic("motion_model", tt.dot(design_matrix, weights)[:, 0])
+            pm.Deterministic("weights", weights)
+
+            # Likelihood to optimize
+            pm.Potential("obs", model.gp.log_likelihood(lc_flux - motion_model))
+
+            # Add deterministic variables for easy of use
+            gp_model, gp_model_var = model.gp.predict(return_var=True)
+            pm.Deterministic("gp_model", gp_model)
+            pm.Deterministic("gp_model_std", np.sqrt(gp_model_var))
+            pm.Deterministic("corrected_flux", lc_flux - motion_model + tt.mean(motion_model))
+            pm.Deterministic("corrected_flux_without_star", lc_flux - motion_model - gp_model + tt.mean(motion_model))
+
+        return model
+
+    def optimize(self, model=None, start=None, **kwargs):
+        """Returns the maximum likelihood solution.
+
+        Parameters
+        ----------
+        model : pymc3.model.Model
+            A pymc3 model
+        start : dict
+            MAP Solution from exoplanet
+        **kwargs : dict
+            Dictionary of arguments to be passed to
+            `lightkurve.correctors.PLDCorrector.create_pymc_model`.
+
+        Returns
+        -------
+        map_soln : dict
+            Maximum likelihood values.
+        """
+        if model is None:
+            model = self.create_pymc_model(**kwargs)
+        if start is None:
+            start = model.test_point
+
+        with model:
+            map_soln = xo.optimize(start=start, vars=[model.logsigma])
+            map_soln = xo.optimize(start=start, vars=[model.logrho, model.logsigma])
+            map_soln = xo.optimize(start=start, vars=[model.logsigma])
+            map_soln = xo.optimize(start=start, vars=[model.logrho, model.logsigma])
+            map_soln = xo.optimize(start=map_soln, vars=[model.logs2])
+            map_soln = xo.optimize(start=map_soln, vars=[model.logrho, model.logsigma, model.logs2])
+
+        self.most_recent_solution = map_soln
+        return map_soln
+
+    def sample(self, model=None, start=None, draws=1000):
+        """Sample the systematics correction model.
+
+        Parameters
+        ----------
+        model : pymc3.model.Model
+            A pymc3 model
+        start : dict
+            MAP Solution from exoplanet
+        draws : int
+            Number of samples
+
+        Returns
+        -------
+        trace : `pymc3.trace`
+            Trace object containing parameters and their samples
+        """
+        # Create the model
+        if model is None:
+            model = self.create_pymc_model()
+        if start is None:
+            start = self.optimize()
+
+        # Initialize the sampler
+        sampler = xo.PyMC3Sampler()
+        with model:
+            # Burn in the sampler
+            burnin = sampler.tune(tune=np.max([int(draws*0.3), 150]),
+                                  start=start,
+                                  step_kwargs=dict(target_accept=0.9),
+                                  chains=4)
+            # Sample the parameters
+            trace = sampler.sample(draws=draws, chains=4)
+
+        return trace
+
+    def correct(self, remove_gp_trend=False, sample=False, draws=1000, **kwargs):
+        """Returns a systematics-corrected light curve.
+
+        Parameters
+        ----------
+        remove_gp_trend : boolean
+            `True` will subtract the fit GP stellar signal from the returned
+            flux light curve.
+        sample : boolean
+            Boolean expression: `True` will sample the output of the optimization
+            step and include robust errors on the output light curve.
+        draws : int
+            Number of samples
+
+        Returns
+        -------
+        corrected_lc : `~lightkurve.lightcurve.LightCurve`
+            Motion noise corrected light curve object.
+        """
+        if remove_gp_trend:
+            corrected_lc = self._get_lightcurve(sample=sample, column_name='corrected_flux_without_star', **kwargs)
         else:
-            # compute the coefficients C on the basis vectors;
-            # the PLD design matrix will be dotted with C to solve for the noise model.
-            ivar = 1.0 / M(rawflux_err)**2 # inverse variance
-            A = np.dot(MX.T, MX * ivar[:, None])
-            B = np.dot(MX.T, M(rawflux) * ivar)
+            corrected_lc = self._get_lightcurve(sample=sample, column_name='corrected_flux', **kwargs)
 
-        # apply prior to design matrix weights for numerical stability
-        A[np.diag_indices_from(A)] += 1e-8
-        C = np.linalg.solve(A, B)
-
-        # compute detrended light curve
-        model = np.dot(X, C)
-        self.detrended_flux = rawflux - (model - np.nanmean(model))
-
-        # Create and return a new LightCurve object with the corrected flux
-        corrected_lc = lc.copy()[nanmask]
-        corrected_lc.flux = self.detrended_flux
-        corrected_lc.flux_err = flux_err
         return corrected_lc
+
+    def _get_lightcurve(self, solution_or_trace=None, sample=False, column_name='corrected_flux', **kwargs):
+        """Helper function to generate light curve objects from a trace."""
+        lc = self.raw_lc.copy()
+        if solution_or_trace is None:
+            if sample:
+                model = self.create_pymc_model(**kwargs)
+                sol = self.optimize(model=model, **kwargs)
+                solution_or_trace = self.sample(model=model, start=sol, draws=draws)
+            else:
+                solution_or_trace = self.optimize(**kwargs)
+        # If a trace is given, find the mean and std
+        if isinstance(solution_or_trace, pm.backends.base.MultiTrace):
+            lc.flux = np.nanmean(solution_or_trace[column_name], axis=0)
+            lc.flux_err = np.nanstd(solution_or_trace[column_name], axis=0)
+        # Otherwise, read value from the dictionary
+        else:
+            lc.flux = solution_or_trace[column_name]
+
+        return lc
+
+    def get_diagnostic_lightcurves(self, solution_or_trace=None, **kwargs):
+        """Return useful diagnostic light curves.
+
+        Parameters
+        ----------
+        solution : dict
+            Dictionary containing output of `exoplanet` optimization.
+        **kwargs : dict
+            Dictionary of arguments to be passed to
+            `lightkurve.correctors.PLDCorrector.optimize`.
+
+        Returns
+        -------
+        raw_lc : `~lightkurve.lightcurve.LightCurve`
+            Light curve object with raw (uncorrected) flux.
+        corrected_lc : `~lightkurve.lightcurve.LightCurve`
+            Motion noise corrected light curve object.
+        gp_lc : `~lightkurve.lightcurve.LightCurve`
+            Light curve object containing GP model of the stellar signal.
+        motion_lc : `~lightkurve.lightcurve.LightCurve`
+            Light curve object with the motion model removed by the corrector.
+        """
+        if self.most_recent_trace is not None and solution_or_trace is None:
+            solution_or_trace = self.most_recent_trace
+        elif self.most_recent_solution is not None and solution_or_trace is None:
+            solution_or_trace = self.most_recent_solution
+
+        corrected_lc = self._get_lightcurve(solution_or_trace=solution_or_trace,
+                                            column_name='corrected_flux', **kwargs)
+        gp_lc = self._get_lightcurve(solution_or_trace=solution_or_trace,
+                                     column_name='gp_model', **kwargs)
+        motion_lc = self._get_lightcurve(solution_or_trace=solution_or_trace,
+                                         column_name='motion_model', **kwargs)
+
+        return self.raw_lc, corrected_lc, gp_lc, motion_lc
+
+    def plot_diagnostics(self, solution=None, **kwargs):
+        """Plots a series of useful figures to help understand the noise removal
+        process.
+
+        Parameters
+        ----------
+        solution : dict
+            Dictionary containing output of `exoplanet` optimization.
+        **kwargs : dict
+            Dictionary of arguments to be passed to
+            `lightkurve.correctors.PLDCorrector.optimize`.
+
+        Returns
+        -------
+        ax : matplotlib.axes._subplots.AxesSubplot
+            The matplotlib axes object.
+        """
+
+        # Generate diagnostic light curves
+        raw_lc, corrected_lc, gp_lc, motion_lc = self.get_diagnostic_lightcurves(solution, **kwargs)
+
+        fig, ax = plt.subplots(3, figsize=(8.45, 10))
+        # Plot the corrected light curve over the raw flux
+        raw_lc.scatter(c='r', alpha=0.3, ax=ax[0], label='Raw Flux')
+        corrected_lc.scatter(c='k', ax=ax[0], label='Corrected Flux')
+        # Plot the stellar model over the raw flux, indicating masked cadences
+        raw_lc.scatter(c='r', alpha=0.3, ax=ax[1], label='Raw Flux')
+        gp_lc.plot(c='k', ax=ax[1], label='Stellar Model')
+        gp_lc[~self.cadence_mask].scatter(ax=ax[1], label='Masked Cadences', marker='d')
+        # Plot the motion model over the raw light curve
+        raw_lc.scatter(c='r', alpha=0.3, ax=ax[2], label='Raw Flux')
+        motion_lc.scatter(c='k', ax=ax[2], label='Noise Model')
+
+        return ax
+
+    def plot_design_matrix(self, design_matrix=None, **kwargs):
+        """Plots the design matrix.
+
+        Parameters
+        ----------
+        design_matrix : np.ndarray
+            Matrix of shape (n_cadences, n_regressors) used to create the
+            motion model.
+        **kwargs : dict
+            Dictionary of arguments to be passed to
+            `lightkurve.correctors.PLDCorrector.create_design_matrix`.
+
+        Returns
+        -------
+        ax : matplotlib.axes._subplots.AxesSubplot
+            The matplotlib axes object.
+        """
+        if design_matrix is None:
+            design_matrix = self.create_design_matrix(**kwargs)
+
+        with plt.style.context(MPLSTYLE):
+            fig, ax = plt.subplots(1, figsize=(8.45, 8.45))
+            ax.imshow(design_matrix, aspect='auto')
+            ax.set_ylabel('Cadence Number')
+            ax.set_xlabel('Regressors')
+
+        return ax
